move `RetrieveFileAtPath` to src/interactors/mod
Also add functions to both the `github` and `gitlab` modules for
constructing the Uri to the manifest
